don't persist categories for favorites
