+++ novawallet/Common/Storage/UserDataModel.xcdatamodeld/MultiassetUserDataModel14.xcdatamodel/contents
@@ -26,7 +26,6 @@
         <attribute name="url" optional="YES" attributeType="URI"/>
     </entity>
     <entity name="CDDAppFavorite" representedClassName="CDDAppFavorite" syncable="YES" codeGenerationType="class">
-        <attribute name="categories" optional="YES" attributeType="String"/>
         <attribute name="icon" optional="YES" attributeType="String"/>
         <attribute name="identifier" optional="YES" attributeType="String"/>
         <attribute name="index" optional="YES" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES"/>
