+++ novawallet/Modules/DApp/DAppList/DAppListInteractor.swift
@@ -82,7 +82,6 @@ final class DAppListInteractor {
             identifier: dApp.url.absoluteString,
             label: dApp.name,
             icon: dApp.icon?.absoluteString,
-            categories: dApp.categories,
             index: index
         )
