+++ novawallet/Modules/DApp/DAppList/ViewModel/DAppListViewModelFactory/DAppListViewModelFactoryProtocol.swift
@@ -1,0 +1,41 @@
+import Foundation
+import SubstrateSdk
+
+protocol DAppListViewModelFactoryProtocol {
+    func createFavoriteDAppName(from model: DAppFavorite) -> String
+
+    func createFavoriteDApps(
+        from list: [DAppFavorite],
+        dAppList: DAppList
+    ) -> [DAppViewModel]
+
+    func createDApps(
+        from category: String?,
+        query: String?,
+        dAppList: DAppList,
+        favorites: [String: DAppFavorite]
+    ) -> DAppListViewModel
+
+    func createDAppSections(
+        from dAppList: DAppList,
+        favorites: [String: DAppFavorite],
+        wallet: MetaAccountModel,
+        hasWalletsListUpdates: Bool,
+        locale: Locale
+    ) -> [DAppListSectionViewModel]
+}
+
+extension DAppListViewModelFactoryProtocol {
+    func createDApps(
+        from category: String?,
+        dAppList: DAppList,
+        favorites: [String: DAppFavorite]
+    ) -> DAppListViewModel {
+        createDApps(
+            from: category,
+            query: nil,
+            dAppList: dAppList,
+            favorites: favorites
+        )
+    }
+}
